feat: #3353 refactor to make more testable
