+++ packages/payments-client/jest.config.js
@@ -4,18 +4,14 @@ const { compilerOptions } = require('./tsconfig')
 
 module.exports = {
   ...baseConfig,
-  testPathIgnorePatterns: [
-    '<rootDir>/src/tests/',
-    '<rootDir>/src/stripe-components/',
-    '<rootDir>/src/stripe-payments-api/',
-  ],
+  testPathIgnorePatterns: ['<rootDir>/src/tests/'],
   moduleNameMapper: {
     ...baseConfig.moduleNameMapper,
     ...pathsToModuleNameMapper(compilerOptions.paths, {
       prefix: '<rootDir>/',
     }),
   },
-  coveragePathIgnorePatterns: ['<rootDir>/src/services', '<rootDir>/src/tests'],
+  coveragePathIgnorePatterns: ['<rootDir>/src/tests', '<rootDir>/src/core/index.tsx'],
   coverageThreshold: {
     global: {
       branches: 0,
