+++ packages/payments-client/src/components/ui/__stubs__/property.ts
@@ -1,6 +1,6 @@
 import { PropertyModel } from '@reapit/foundations-ts-definitions'
 
-export const data: PropertyModel = {
+export const stubPropertyModel: PropertyModel = {
   id: 'OXF200008',
   created: '2018-12-12T12:30:23.0000000Z',
   modified: '2019-01-08T12:30:34.0000000Z',
