+++ packages/payments-client/src/components/ui/__tests__/payment-custom-section.test.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
-import PaymentCustomerSection from '../../payments/payment-customer-section'
-import { data } from '../../__stubs__/payment'
+import PaymentCustomerSection from '../payment-customer-section'
+import { stubPaymentModel } from '../__stubs__/payment'
 
 describe('PaymentCustomerSection', () => {
   it('should match a snapshot', () => {
-    const customer = data.customer
+    const customer = stubPaymentModel.customer
     const wrapper = shallow(<PaymentCustomerSection customer={customer} />)
     expect(wrapper).toMatchSnapshot()
   })
