+++ packages/payments-client/src/components/ui/__tests__/property-section.test.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
-import PropertySection from '../../payments/property-section'
-import { data } from '../../__stubs__/property'
+import PropertySection from '../property-section'
+import { stubPropertyModel } from '../__stubs__/property'
 
 describe('PropertySection', () => {
   it('should match a snapshot', () => {
-    const wrapper = shallow(<PropertySection property={data} />)
+    const wrapper = shallow(<PropertySection property={stubPropertyModel} />)
     expect(wrapper).toMatchSnapshot()
   })
 })
