+++ packages/payments-client/src/components/ui/property-section.tsx
