+++ packages/payments-client/src/constants/routes.ts
@@ -1,8 +1,6 @@
-const Routes = {
+export const Routes = {
   HOME: '/',
   LOGIN: '/login',
   PAYMENTS: '/payments',
   PAYMENT: '/payments/:paymentId',
 }
-
-export default Routes
