--- packages/payments-client/src/constants/toast-messages.ts
@@ -1,5 +1,0 @@
-export const toastMessages = {
-  UPDATE_PAYMENT_STATUS_POSTED: 'Payment is posted',
-  UPDATE_PAYMENT_STATUS_REJECTED: 'Payment failed',
-  FAILED_TO_UPDATE_PAYMENT_STATUS: 'Failed to update the payment status',
-}
