+++ packages/payments-client/src/core/router.tsx
@@ -2,8 +2,7 @@ import * as React from 'react'
 import { Route, Router as BrowserRouter, Switch, Redirect } from 'react-router-dom'
 import { createBrowserHistory } from 'history'
 import { SWRConfig } from 'swr'
-
-import Routes from '../constants/routes'
+import { Routes } from '../constants/routes'
 import PrivateRouteWrapper from './private-route-wrapper'
 import { platformFetcher } from '../utils/fetcher'
