+++ packages/payments-client/src/services/__stubs__/payment.ts
@@ -1,4 +1,4 @@
-import { UpdateStatusBody, UpdateStatusParams } from '../payment'
+import { UpdateStatusBody, UpdateStatusParams } from '../../types/payment'
 
 export const body: UpdateStatusBody = { status: 'posted', externalReference: 'SOME_POSTED_REF' }
 export const params: UpdateStatusParams = {
